jwt authentication added   , authorization also added

+++ app/Http/Controllers/UsersController.php
@@ -6,9 +6,14 @@ use Illuminate\Http\Request;
 use App\User;
 use App\Http\Resources\UserResource;
 use Illuminate\Support\Facades\Auth;
+use JWTAuth;
+use Tymon\JWTAuth\Exceptions\JWTException;
+// use Auth;
 
 class UsersController extends Controller
+
 {
+
     public function index()
     {
         $user = User::paginate(5);
@@ -19,53 +24,40 @@ class UsersController extends Controller
     public function show($id)
     {
         // get user
-        $user = User::findOrfail($id);
-
-        // return user as a resource
-        return  new UserResource($user);
+       if(auth()->user()->id == $id)
+            return  new UserResource($user);
+        else
+            return response(['message' => 'you can only view your own info'] , 401);
     }
 
-    public function checkLogin(Request $request)
-    {
-        if (Auth::attempt(['name' => $request->username, 'password' => $request->password])){
+    
 
-            return response()->json([
-                'login' => true,
-                'username'=> Auth::user()->name
-            ],200);
-
-        }
-        else{
-            return response()->json([
-                'data' => 'invalid login details'
-            ],500);
-
-        }
-
-    }
-
-    public function store(Request $request)
-    {
-        $user = new User();
-
-        $user->name = $request->name;
-        $user->email = $request->email;
-        $user->password = $request->password ? $request->password: '12345' ;
-
-        if($user->save()){
-            return new UserResource($user);
-        }
-    }
+    
 
     public function update(Request $request , $id)
     {
         $user = User::findorfail($id);
 
-        $user->name = $request->name;
-        $user->email = $request->email;
+        if(auth()->user()->id  == $id)
+        {
+            $validator = Validator::make($request->all() ,[
+                    'name'=>'required',
+                    'email'=>'required|email|unique:users',
 
-        if($user->save()){
-            return new UserResource($user);
+                ]);
+            if($validator -> fails())
+            {
+                return response(['errors' , $validator->errors()->all()]  , 400);
+            }
+            $user->name = $request->name;
+            $user->email = $request->email;
+
+            $user->save();
+
+        }
+        else
+        {
+            return response(['message' => 'you can only modify your own info'] , 401);
         }
     }
 
@@ -74,8 +66,15 @@ class UsersController extends Controller
         // get user
         $user = User::findOrfail($id);
 
-       if($user->delete()){
-        return  new UserResource($user);
-       }
+       if(auth()->user()->id  == $id)
+        {
+            if($user->delete()){
+                return  new UserResource($user);
+            }
+        }
+        else
+        {
+          return response(['message' => 'you can only delete your own info'] , 401);
+        }
     }
 }
